requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
# add test line

attrs==22.1.0
azure-identity
azure-keyvault-secrets
requests
python-json-logger
structlog~=22.3.0
pytz
